Change the link URL for creating a new redirect
Problem:

Users need to be familiar with Github's forking/pull request flow
in order to add a redirect to the list.
This is a barrier for entry for non-technical users.

Solution:

Take advantage of Github's excellent URL schema
to point users at the exact page
where they can add redirects.

The new URL is:
https://github.com/codeforamerica/redirect/edit/master/redirects.json#L2

If this is the user's first time editing the file,
they will be prompted to fork the repository
in order to make changes to the file.

After that, it will redirect the user
to the "edit" page for the file.
They can edit the JSON in their browser,
and click "Propose file change" and "Create pull request"
to submit it for review.